+++ R/bru.inference.R
@@ -799,6 +799,15 @@ like <- function(formula = . ~ ., family = "gaussian", data = NULL,
     # or similar.
     # For Spatial models, keep the old behaviour for backwards compatibility for
     # now, but can likely realign that in the future after more testing.
+    # Save general response data to add to response (precomputed covariates etc)
+    if (!is.null(response_data)) {
+      data_ <- response_data
+    } else {
+      data_ <- data
+    }
+    if (inherits(data_, "Spatial")) {
+      data_ <- as.data.frame(data_)
+    }
     if (ips_is_Spatial) {
       if ("coordinates" %in% names(response)) {
         idx <- names(response) %in% "coordinates"
@@ -820,6 +829,14 @@ like <- function(formula = . ~ ., family = "gaussian", data = NULL,
       response_data <- NULL
       N_data <- NROW(data)
     }
+
+    # Add back additional data
+    additional_data_names <- setdiff(names(data_), names(data))
+    if ((length(additional_data_names) > 0) &&
+        (NROW(data_) == N_data)) {
+      data <- cbind(data, data_[additional_data_names])
+    }
+
     if (ips_is_Spatial) {
       ips <- as.data.frame(ips)
     } else {
@@ -827,7 +844,7 @@ like <- function(formula = . ~ ., family = "gaussian", data = NULL,
         sf::st_geometry(ips) <- "geometry"
       }
     }
-    dim_names <- intersect(names(data), names(ips))
+
     if (identical(options[["bru_compress_cp"]], TRUE)) {
       allow_combine <- TRUE
       response_data <- data.frame(
@@ -853,9 +870,9 @@ like <- function(formula = . ~ ., family = "gaussian", data = NULL,
         )
       }
       expr <- parse(text = expr_text)
-      data <- rbind(
-        cbind(data[dim_names], BRU_aggregate = TRUE),
-        cbind(ips[dim_names], BRU_aggregate = FALSE)
+      data <- dplyr::bind_rows(
+        cbind(data, BRU_aggregate = TRUE),
+        cbind(ips, BRU_aggregate = FALSE)
       )
       formula
     } else {
@@ -869,10 +886,7 @@ like <- function(formula = . ~ ., family = "gaussian", data = NULL,
           rep(0, NROW(ips))
         )
       )
-      data <- rbind(
-        data[dim_names],
-        ips[dim_names]
-      )
+      data <- dplyr::bind_rows(data, ips)
     }
     if (ips_is_Spatial) {
       non_coordnames <- setdiff(names(data), data_coordnames)
@@ -2338,15 +2352,39 @@ iinla <- function(model, lhoods, initial = NULL, options) {
     original_timings <- old.result[["bru_iinla"]][["timings"]]
   }
 
+  bru_log_message(
+    "iinla: Evaluate component inputs",
+    verbose = options$bru_verbose,
+    verbose_store = options$bru_verbose_store,
+    verbosity = 3
+  )
   inputs <- evaluate_inputs(model, lhoods = lhoods, inla_f = TRUE)
+  bru_log_message(
+    "iinla: Evaluate component linearisations",
+    verbose = options$bru_verbose,
+    verbose_store = options$bru_verbose_store,
+    verbosity = 3
+  )
   comp_lin <- evaluate_comp_lin(model,
     input = inputs,
     state = states[[length(states)]],
     inla_f = TRUE
   )
+  bru_log_message(
+    "iinla: Evaluate component simplifications",
+    verbose = options$bru_verbose,
+    verbose_store = options$bru_verbose_store,
+    verbosity = 3
+  )
   comp_simple <- evaluate_comp_simple(model,
     input = inputs,
     inla_f = TRUE
+  )
+  bru_log_message(
+    "iinla: Evaluate predictor linearisation",
+    verbose = options$bru_verbose,
+    verbose_store = options$bru_verbose_store,
+    verbosity = 3
   )
   lin <- bru_compute_linearisation(
     model,
@@ -2361,12 +2399,24 @@ iinla <- function(model, lhoods, initial = NULL, options) {
     # Always compute linearisation (above)
   }
 
+  bru_log_message(
+    "iinla: Construct inla stack",
+    verbose = options$bru_verbose,
+    verbose_store = options$bru_verbose_store,
+    verbosity = 3
+  )
   # Initial stack
   idx <- evaluate_index(model, lhoods)
   stk <- bru_make_stack(lhoods, lin, idx)
 
   stk.data <- INLA::inla.stack.data(stk)
   inla.options$control.predictor$A <- INLA::inla.stack.A(stk)
+  bru_log_message(
+    "iinla: Model initialisation completed",
+    verbose = options$bru_verbose,
+    verbose_store = options$bru_verbose_store,
+    verbosity = 3
+  )
 
   k <- 1
   interrupt <- FALSE
@@ -2581,10 +2631,22 @@ iinla <- function(model, lhoods, initial = NULL, options) {
           )
           state <- line_search[["state"]]
         }
+        bru_log_message(
+          "iinla: Evaluate component linearisations",
+          verbose = options$bru_verbose,
+          verbose_store = options$bru_verbose_store,
+          verbosity = 3
+        )
         comp_lin <- evaluate_comp_lin(model,
           input = inputs,
           state = state,
           inla_f = TRUE
+        )
+        bru_log_message(
+          "iinla: Evaluate predictor linearisation",
+          verbose = options$bru_verbose,
+          verbose_store = options$bru_verbose_store,
+          verbosity = 3
         )
         lin <- bru_compute_linearisation(
           model,